Fix C++20 deprecated-enum-float-conversion warnings
C++20 deprecates implicit conversion between an unscoped enum and a
floating-point type in arithmetic/comparisons (P1120). This codebase
has several old-style "enum as named integer constant" definitions
(NAV::BIAS_DANGER, SC_MEDIUM_RADIUS, Z_CULL_OFFSET, RANGE_NULL,
shaderSort_t values, etc.) that get mixed into float expressions
throughout the navigation code and renderers.

Add explicit (float) casts at each site to preserve the exact prior
implicit-conversion behavior. Purely a warning fix - no logic change.

+++ code/Ratl/grid_vs.h
@@ -211,11 +211,11 @@ public:
 		float	point[2] = {xReal, yReal};
 		for (int i=0; i<2; i++)
 		{
-			if (point[i]<mMins[i] || mMins[i]==RANGE_NULL)
+			if (point[i]<mMins[i] || mMins[i]==(float)RANGE_NULL)
 			{
 				mMins[i] = point[i];
 			}
-			if (point[i]>mMaxs[i] || mMaxs[i]==RANGE_NULL)
+			if (point[i]>mMaxs[i] || mMaxs[i]==(float)RANGE_NULL)
 			{
 				mMaxs[i] = point[i];
 			}
