+++ code/rd-vanilla/tr_shader.cpp
@@ -2800,7 +2800,7 @@ static shader_t *GeneratePermanentShader( void ) {
 
 	*newShader = shader;
 
-	if ( shader.sort <= /*SS_OPAQUE*/SS_SEE_THROUGH ) {
+	if ( shader.sort <= /*SS_OPAQUE*/(float)SS_SEE_THROUGH ) {
 		newShader->fogPass = FP_EQUAL;
 	} else if ( shader.contentFlags & CONTENTS_FOG ) {
 		newShader->fogPass = FP_LE;
@@ -2865,7 +2865,7 @@ static int VertexLightingCollapse( void ) {
 	int		finalstagenum=1;
 
 	// if we aren't opaque, just use the first pass
-	if ( shader.sort == SS_OPAQUE ) {
+	if ( shader.sort == (float)SS_OPAQUE ) {
 
 		// pick the best texture for the single pass
 		bestStage = &stages[0];
